refactor(app): extract boot/hydration/idle-preload orchestration into useAppBootstrap
- Move the batched boot payload, crash recovery gate, hydration,
  post-hydration housekeeping hooks (shortcut hints, getting-started,
  orchestration milestones, agent-waiting nudge, forge recommendation),
  plugin deep-link handling, notification/OS-DND hydration, and the
  idle-time palette/dialog chunk prefetch out of AppInner
- AppInner now composes useAppBootstrap() + useModalResetKeys() in a
  fixed unconditional order — hook call sequence is unchanged
- No logic changes
