+++ src/hooks/app/index.ts
@@ -3,6 +3,7 @@ export type { SafeBootResult } from "./useAppBoot";
 export { useE2EBridges } from "./useE2EBridges";
 export { useModalResetKeys } from "./useModalResetKeys";
 export type { ModalResetKeys } from "./useModalResetKeys";
+export { useAppBootstrap } from "./useAppBootstrap";
 export { useAppHydration } from "./useAppHydration";
 export { useShortcutHints } from "./useShortcutHints";
 export { usePanelStoreBootstrap } from "./usePanelStoreBootstrap";
